LOTS of changes. Moved to vundle, away from Janus.
It took quite a bit but I think it was worth it... Hopefully this
restores all the major things I liked about Janus. For Nerdtree, I
needed the stuff in /plugins/Nerdtree to make it work the same way. Most
other things are in vimrc and gvimrc.

Also, fixed backup, and undo issues.